Adds the postgresql-data-k8s charm to the bundle
The postgresql-data-k8s charm can relate to the postgresql-k8s
charm's db-admin relation, which will allow it to update the
configured database. This can be used to inject a public dataset,
given through the sql-dump-url config option, and can be reset to
that dump periodically.

This can be used for a public instance of Waltz, which contains
a public dataset.