Fixes image / bundle publishing jobs
Dispatch create_release.yaml on a new release. The create_release.yaml action will
create a new branch based on that release, create a commit which will update the
bundle.yaml file to make the finos-waltz-k8s charm point towards that release,
and push the new bundle to Charmhub.

When there's a new Waltz image release, we do a charmcraft release
specific for that version in Charmhub.

We're manually packing and uploading the bundle instead of using the
upload-bundle action because of an issue [1]. This will have to be updated
later, when the issue is resolved.

[1] https://github.com/canonical/charming-actions/issues/40