Update publishing documentation
Updates the documentation with an increased time-to-live for the
charmcraft login credentials. The new TTL is 6 months replacing the old
value of 18 days. This attenuates certificate expiration issues.

Adds `gh secret` command so that we can set the charmcraft secret
without going through the Github Settings UI.

Adds reference to the error message thrown when credentials are expired
for charmcraft.